Made logdata and selog more consistent
