+++ nexus-writer/src/nexus/hdf5_file/mod.rs
@@ -5,8 +5,7 @@ mod run_file_components;
 use run_file_components::EventRun;
 
 use hdf5_writer::{
-    add_attribute_to, add_new_group_to, create_resizable_2d_dataset,
-    create_resizable_2d_dataset_dyn_type, create_resizable_dataset, set_group_nx_class,
-    set_slice_to, set_string_to,
+    add_attribute_to, add_new_group_to, create_resizable_dataset, set_group_nx_class, set_slice_to,
+    set_string_to,
 };
-pub(crate) use run_file::{RunFile, VarArrayTypeSettings};
+pub(crate) use run_file::RunFile;
