+++ nexus-writer/src/nexus/hdf5_file/run_file_components/mod.rs
@@ -3,10 +3,7 @@ mod runlog_file;
 mod selog_file;
 mod timeseries_file;
 
-use super::{
-    add_new_group_to, create_resizable_2d_dataset, create_resizable_2d_dataset_dyn_type,
-    create_resizable_dataset,
-};
+use super::add_new_group_to;
 
 pub(crate) use event_run_file::EventRun;
 pub(crate) use runlog_file::RunLog;
