+++ nexus-writer/src/nexus/hdf5_file/run_file_components/selog_file.rs
@@ -1,393 +1,76 @@
-use super::{
-    add_new_group_to, create_resizable_2d_dataset, create_resizable_2d_dataset_dyn_type,
-    create_resizable_dataset,
-    timeseries_file::{Slice1D, TimeSeriesEntry, TimeSeriesOwner},
+use super::{add_new_group_to, timeseries_file::TimeSeriesDataSource};
+use crate::nexus::{
+    hdf5_file::{
+        hdf5_writer::create_resizable_dataset,
+        run_file_components::timeseries_file::get_dataset_builder,
+    },
+    nexus_class as NX,
 };
-use crate::nexus::{nexus_class as NX, NexusSettings};
-use anyhow::{anyhow, Result};
-use hdf5::{
-    types::{FloatSize, IntSize, TypeDescriptor},
-    Dataset, Group, H5Type,
-};
-use ndarray::s;
-use supermusr_streaming_types::{
-    ecs_se00_data_generated::{se00_SampleEnvironmentData, ValueUnion},
-    flatbuffers::{Follow, Vector},
-};
+use anyhow::Result;
+use hdf5::{Group, SimpleExtents};
+use std::fmt::Debug;
+use supermusr_streaming_types::ecs_se00_data_generated::se00_SampleEnvironmentData;
 use tracing::debug;
 
 #[derive(Debug)]
 pub(crate) struct SeLog {
-    /*num_selogs: usize,
-
-    channel: Dataset,
-    packet_timestamp: Dataset,
-    time_delta: Dataset,
-    timestamp_location: Dataset,
-    values: Dataset,
-    timestamps: Dataset,
-    message_counter: Dataset,*/
     parent: Group,
 }
+
 impl SeLog {
     #[tracing::instrument]
-    pub(crate) fn new_selog(parent: &Group, settings: &NexusSettings) -> Result<Self> {
-        /*
-        let selog = add_new_group_to(parent, "selog", NX::SELOG)?;
-        let channel = create_resizable_dataset::<i32>(&selog, "channel", 0, 32)?;
-        let packet_timestamp = create_resizable_dataset::<i64>(&selog, "packet_timestamp", 0, 32)?;
-        let time_delta = create_resizable_dataset::<f64>(&selog, "time_delta", 0, 32)?;
-        let timestamp_location =
-            create_resizable_dataset::<u8>(&selog, "timestamp_location", 0, 32)?;
-
-        let values = create_resizable_2d_dataset_dyn_type(
-            &selog,
-            "values",
-            &settings.sample_env.data_type,
-            (0, 0),
-            (32, 32),
-        )?;
-        let timestamps =
-            create_resizable_2d_dataset::<i64>(&selog, "timestamps", (0, 0), (32, 32))?;
-
-        let message_counter = create_resizable_dataset::<i64>(&selog, "message_counter", 0, 32)?;
-
-        Ok(Self {
-            num_selogs: 0,
-            channel,
-            packet_timestamp,
-            time_delta,
-            timestamp_location,
-            values,
-            timestamps,
-            message_counter,
-        }) */
-        let selogs = add_new_group_to(parent, "selog", NX::SELOG)?;
-        Ok(Self { parent: selogs })
+    pub(crate) fn new_selog(parent: &Group) -> Result<Self> {
+        let logs = add_new_group_to(parent, "selog", NX::RUNLOG)?;
+        Ok(Self { parent: logs })
     }
 
     #[tracing::instrument]
-    pub(crate) fn open_from_selog_group(parent: &Group) -> Result<Self> {
-        /*let selog = parent.group("selog")?;
-        let channel = selog.dataset("channel")?;
-        let packet_timestamp = selog.dataset("packet_timestamp")?;
-        let time_delta = selog.dataset("time_delta")?;
-        let timestamp_location = selog.dataset("timestamp_location")?;
-        let values = selog.dataset("values")?;
-        let timestamps = selog.dataset("timestamps")?;
-        let message_counter = selog.dataset("message_counter")?;
-
-        Ok(Self {
-            num_selogs: 0,
-            channel,
-            packet_timestamp,
-            time_delta,
-            timestamp_location,
-            values,
-            timestamps,
-            message_counter,
-        })*/
+    pub(crate) fn open_selog(parent: &Group) -> Result<Self> {
         let parent = parent.group("selog")?;
         Ok(Self { parent })
     }
 
-    fn get_hdf5_type(fb_type: ValueUnion) -> Result<TypeDescriptor> {
-        let datatype = match fb_type {
-            ValueUnion::Int8Array => TypeDescriptor::Integer(IntSize::U1),
-            ValueUnion::Int16Array => TypeDescriptor::Integer(IntSize::U2),
-            ValueUnion::Int32Array => TypeDescriptor::Integer(IntSize::U4),
-            ValueUnion::Int64Array => TypeDescriptor::Integer(IntSize::U8),
-            ValueUnion::UInt8Array => TypeDescriptor::Unsigned(IntSize::U1),
-            ValueUnion::UInt16Array => TypeDescriptor::Unsigned(IntSize::U2),
-            ValueUnion::UInt32Array => TypeDescriptor::Unsigned(IntSize::U4),
-            ValueUnion::UInt64Array => TypeDescriptor::Unsigned(IntSize::U8),
-            ValueUnion::FloatArray => TypeDescriptor::Float(FloatSize::U4),
-            ValueUnion::DoubleArray => TypeDescriptor::Float(FloatSize::U8),
-            t => {
-                return Err(anyhow!(
-                    "Invalid flatbuffers selogdata type {}",
-                    t.variant_name().unwrap()
-                ))
-            }
-        };
-        Ok(datatype)
-    }
-
-    #[tracing::instrument(skip(self), fields(message_number, num_events))]
-    pub(crate) fn push_selogdata(
+    #[tracing::instrument(skip(self))]
+    fn create_new_selogdata_in_selog(
         &mut self,
-        selogdata: se00_SampleEnvironmentData,
-        settings: &NexusSettings,
+        selog: &se00_SampleEnvironmentData,
+    ) -> Result<Group> {
+        add_new_group_to(&self.parent, selog.name(), NX::SELOG).and_then(|parent_group| {
+            let group = add_new_group_to(&parent_group, "value log", NX::SELOG)?;
+            create_resizable_dataset::<i32>(&group, "time", 0, 1024)?;
+            get_dataset_builder(&selog.get_hdf5_type()?, &group)?
+                .shape(SimpleExtents::resizable(vec![0]))
+                .chunk(1024)
+                .create("value")?;
+            Ok::<_, anyhow::Error>(parent_group)
+        })
+    }
+
+    #[tracing::instrument(skip(self))]
+    pub(crate) fn push_selogdata_to_selog(
+        &mut self,
+        selog: &se00_SampleEnvironmentData,
     ) -> Result<()> {
-        tracing::Span::current().record("selog_number", self.num_selogs);
-        debug!("{:?}", selogdata.values_type());
-        let datatype = Self::get_hdf5_type(selogdata.values_type())?;
-        if datatype != settings.sample_env.data_type {
-            return Err(anyhow!("Incorrect Datatype"));
+        debug!("Type: {0:?}", selog.values_type());
+
+        let timeseries = self
+            .parent
+            .group(selog.name())
+            .or_else(|err| {
+                self.create_new_selogdata_in_selog(selog)
+                    .map_err(|e| e.context(err))
+            })?
+            .group("value log")?;
+        let timestamps = timeseries.dataset("time")?;
+        let values = timeseries.dataset("value")?;
+        let slice = selog.write_values_to_dataset(&values)?;
+        if let Some(source_timestamps) = selog.timestamps() {
+            timestamps.write_slice(
+                source_timestamps.iter().collect::<Vec<_>>().as_slice(),
+                slice,
+            )?;
+        } else if selog.time_delta() > 0.0 {
         }
-
-        let slice = s![self.num_selogs..(self.num_selogs + 1)];
-
-        self.channel.resize(self.num_selogs + 1).unwrap();
-        self.channel.write_slice(&[selogdata.channel()], slice)?;
-
-        self.packet_timestamp.resize(self.num_selogs + 1).unwrap();
-        self.packet_timestamp
-            .write_slice(&[selogdata.packet_timestamp()], slice)?;
-
-        self.time_delta.resize(self.num_selogs + 1).unwrap();
-        self.time_delta
-            .write_slice(&[selogdata.time_delta()], slice)?;
-
-        self.timestamp_location.resize(self.num_selogs + 1).unwrap();
-        self.timestamp_location
-            .write_slice(&[selogdata.timestamp_location().0], slice)?;
-
-        let array_slice = s![
-            self.num_selogs..(self.num_selogs + 1),
-            0..settings.sample_env.array_length
-        ];
-
-        self.timestamps.resize(self.num_selogs + 1).unwrap();
-        self.timestamps.write_slice(
-            &selogdata.timestamps().unwrap().iter().collect::<Vec<_>>(),
-            array_slice,
-        )?;
-
-        self.values
-            .resize((self.num_selogs + 1, settings.sample_env.array_length))
-            .unwrap();
-        match settings.sample_env.data_type {
-            TypeDescriptor::Integer(sz) => match sz {
-                IntSize::U1 => self.values.write_slice(
-                    &selogdata
-                        .values_as_int_8_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                IntSize::U2 => self.values.write_slice(
-                    &selogdata
-                        .values_as_int_16_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                IntSize::U4 => self.values.write_slice(
-                    &selogdata
-                        .values_as_int_32_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                IntSize::U8 => self.values.write_slice(
-                    &selogdata
-                        .values_as_int_64_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-            },
-            TypeDescriptor::Unsigned(sz) => match sz {
-                IntSize::U1 => self.values.write_slice(
-                    &selogdata
-                        .values_as_uint_8_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                IntSize::U2 => self.values.write_slice(
-                    &selogdata
-                        .values_as_uint_16_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                IntSize::U4 => self.values.write_slice(
-                    &selogdata
-                        .values_as_uint_32_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                IntSize::U8 => self.values.write_slice(
-                    &selogdata
-                        .values_as_uint_64_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-            },
-            TypeDescriptor::Float(sz) => match sz {
-                FloatSize::U4 => self.values.write_slice(
-                    &selogdata
-                        .values_as_float_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-                FloatSize::U8 => self.values.write_slice(
-                    &selogdata
-                        .values_as_double_array()
-                        .unwrap()
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    array_slice,
-                ),
-            },
-            _ => {
-                return Err(anyhow!(
-                    "Invalid HDF5 type: {}",
-                    settings.sample_env.data_type.to_string()
-                ))
-            }
-        }?;
-
-        self.message_counter.resize(self.num_selogs + 1).unwrap();
-        self.message_counter
-            .write_slice(&[selogdata.message_counter()], slice)?;
-
-        self.num_selogs += 1;
-
-        Ok(())
-    }
-}
-
-impl<'a> TimeSeriesEntry<'a> for se00_SampleEnvironmentData<'a> {
-    fn timestamp(&self) -> i64 {
-        self.packet_timestamp()
-    }
-}
-
-impl<'a> TimeSeriesOwner<'a> for SeLog {
-    type DataSource = se00_SampleEnvironmentData<'a>;
-
-    #[tracing::instrument]
-    fn write_data_value_to_dataset_slice(
-        type_descriptor: &TypeDescriptor,
-        source: &Self::DataSource,
-        target: &mut Dataset,
-        slice: Slice1D,
-    ) -> Result<()> {
-        let error = anyhow!("Could not convert value to type {type_descriptor:?}");
-        match type_descriptor {
-            TypeDescriptor::Integer(sz) => match sz {
-                IntSize::U1 => target.write_slice(
-                    &source
-                        .values_as_int_8_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                IntSize::U2 => target.write_slice(
-                    &source
-                        .values_as_int_16_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                IntSize::U4 => target.write_slice(
-                    &source
-                        .values_as_int_32_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                IntSize::U8 => target.write_slice(
-                    &source
-                        .values_as_int_64_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-            },
-            TypeDescriptor::Unsigned(sz) => match sz {
-                IntSize::U1 => target.write_slice(
-                    &source
-                        .values_as_uint_8_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                IntSize::U2 => target.write_slice(
-                    &source
-                        .values_as_uint_16_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                IntSize::U4 => target.write_slice(
-                    &source
-                        .values_as_uint_32_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                IntSize::U8 => target.write_slice(
-                    &source
-                        .values_as_uint_64_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-            },
-            TypeDescriptor::Float(sz) => match sz {
-                FloatSize::U4 => target.write_slice(
-                    &source
-                        .values_as_float_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-                FloatSize::U8 => target.write_slice(
-                    &source
-                        .values_as_double_array()
-                        .ok_or(error)?
-                        .value()
-                        .iter()
-                        .collect::<Vec<_>>(),
-                    slice,
-                ),
-            },
-            _ => unreachable!("Unreachable HDF5 TypeDescriptor reached"),
-        }?;
         Ok(())
     }
 }
